--- api/src/main/java/org/pmiops/workbench/db/model/DbUserBase.java
@@ -1,24 +1,0 @@
-package org.pmiops.workbench.db.model;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-// Minimal User representation
-@MappedSuperclass
-public class DbUserBase {
-  private long userId;
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "user_id")
-  public long getUserId() {
-    return userId;
-  }
-
-  public void setUserId(long userId) {
-    this.userId = userId;
-  }
-}
