+++ api/src/test/java/org/pmiops/workbench/db/dao/UserServiceTest.java
@@ -24,17 +24,12 @@ import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
 import org.pmiops.workbench.actionaudit.targetproperties.BypassTimeTargetProperty;
 import org.pmiops.workbench.compliance.ComplianceService;
 import org.pmiops.workbench.config.WorkbenchConfig;
-import org.pmiops.workbench.db.model.DbInstitution;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbUserTermsOfService;
-import org.pmiops.workbench.db.model.DbVerifiedInstitutionalAffiliation;
 import org.pmiops.workbench.exceptions.NotFoundException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.firecloud.model.FirecloudNihStatus;
 import org.pmiops.workbench.google.DirectoryService;
-import org.pmiops.workbench.model.DuaType;
-import org.pmiops.workbench.model.InstitutionalRole;
-import org.pmiops.workbench.model.OrganizationType;
 import org.pmiops.workbench.moodle.ApiException;
 import org.pmiops.workbench.moodle.model.BadgeDetailsV1;
 import org.pmiops.workbench.moodle.model.BadgeDetailsV2;
@@ -62,12 +57,12 @@ public class UserServiceTest {
 
   // An arbitrary timestamp to use as the anchor time for access module test cases.
   private static final Instant START_INSTANT = Instant.parse("2000-01-01T00:00:00.00Z");
+  @Deprecated private static final long TIMESTAMP_MSECS = START_INSTANT.toEpochMilli();
   private static final long TIMESTAMP_SECS = START_INSTANT.getEpochSecond();
   private static final FakeClock PROVIDED_CLOCK = new FakeClock(START_INSTANT);
   private static final int CLOCK_INCREMENT_MILLIS = 1000;
   private static DbUser providedDbUser;
   private static WorkbenchConfig providedWorkbenchConfig;
-  private DbInstitution dbInstitution;
 
   @MockBean private FireCloudService mockFireCloudService;
   @MockBean private ComplianceService mockComplianceService;
@@ -77,8 +72,6 @@ public class UserServiceTest {
 
   @Autowired private UserService userService;
   @Autowired private UserDao userDao;
-  @Autowired private VerifiedInstitutionalAffiliationDao verifiedInstitutionalAffiliationDao;
-  @Autowired private InstitutionDao institutionDao;
 
   @Import(UserServiceTestConfiguration.class)
   @TestConfiguration
@@ -121,29 +114,6 @@ public class UserServiceTest {
     // but it was injecting null clocks giving NPEs long after construction, and so far this
     // is the only working approach I've seen.
     PROVIDED_CLOCK.setInstant(START_INSTANT);
-
-    dbInstitution = buildInstitution();
-    final DbVerifiedInstitutionalAffiliation dbAffiliation = buildAffiliation(dbInstitution);
-  }
-
-  public DbVerifiedInstitutionalAffiliation buildAffiliation(DbInstitution dbInstitution) {
-    final DbVerifiedInstitutionalAffiliation dbAffiliation =
-        new DbVerifiedInstitutionalAffiliation();
-    dbAffiliation.setVerifiedInstitutionalAffiliationId(101L);
-    dbAffiliation.setInstitution(dbInstitution);
-    dbAffiliation.setUser(providedDbUser);
-    dbAffiliation.setInstitutionalRoleEnum(InstitutionalRole.FELLOW);
-    return verifiedInstitutionalAffiliationDao.save(dbAffiliation);
-  }
-
-  public DbInstitution buildInstitution() {
-    final DbInstitution dbInstitution = new DbInstitution();
-    dbInstitution.setDisplayName("Disney World");
-    dbInstitution.setDuaTypeEnum(DuaType.MASTER);
-    dbInstitution.setOrganizationTypeEnum(OrganizationType.OTHER);
-    dbInstitution.setOrganizationTypeOtherText("Happiest place on earth.");
-    dbInstitution.setShortName("disney");
-    return institutionDao.save(dbInstitution);
   }
 
   @Test
@@ -164,7 +134,7 @@ public class UserServiceTest {
     // The user should be updated in the database with a non-empty completion and expiration time.
     DbUser user = userDao.findUserByUsername(USERNAME);
     assertThat(user.getComplianceTrainingCompletionTime())
-        .isEqualTo(new Timestamp(START_INSTANT.toEpochMilli()));
+        .isEqualTo(new Timestamp(TIMESTAMP_MSECS));
     assertThat(user.getComplianceTrainingExpirationTime())
         .isEqualTo(Timestamp.from(Instant.ofEpochMilli(expiry)));
