+++ api/src/test/java/org/pmiops/workbench/profile/ProfileServiceTest.java
@@ -61,7 +61,6 @@ public class ProfileServiceTest {
 
   private static final VerifiedInstitutionalAffiliation BROAD_AFFILIATION =
       new VerifiedInstitutionalAffiliation()
-          .institutionalRoleOtherText("")
           .institutionShortName("Broad")
           .institutionDisplayName("The Broad Institute")
           .institutionalRoleEnum(InstitutionalRole.ADMIN);
